+++ packages/vendure-plugin-sendcloud/src/api/constants.ts
@@ -1,0 +1,2 @@
+export const loggerCtx = 'SendcloudPlugin';
+export const PLUGIN_OPTIONS = Symbol('SENDCLOUD_OPTIONS');