+++ CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.20.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v3.19.0...v3.20.0) (2022-09-28)
+
+### Features
+
+- **goedgepickt:** customfields push ([0966780](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/0966780033eb74d2d93e117cc7543da5d57d84d2))
+
+# [3.19.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v3.18.0...v3.19.0) (2022-09-27)
+
+### Bug Fixes
+
+- sendcloud wip ([ffef3fc](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/ffef3fc0f9e2127c3cb4477dc57a9fb8889919bf))
+
+### Features
+
+- **goedgepickt:** string ([b539b32](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/b539b324be8aabf6357313581720f095cabf5dce))
+- more wip for sendcloud p[lugin ([1270b4a](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/1270b4abe89ea3410e5cc6e44dde6dd0ce5fa742))
+- **sendcloud:** cloud icon ([8bbc390](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/8bbc390bf63da8e6fa79aacef65257f3a9b7268c))
+- **sendcloud:** finished sc impl with e2e ([50eefc5](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/50eefc53c6988166bec8a9d6d233a564dbec893b))
+- **sendcloud:** fulfill with sendcloud ([29c4a71](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/29c4a7137900303152d6f6b63408313edf58e3a8))
+- **sendcloud:** header test ([a478e9f](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/a478e9f7e8b9ef8db97249f3ce61815a6ad4b86c))
+- **sendcloud:** rEADME ([4f7a288](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/4f7a28887ee1bfb612936c1276583e37d99bfc24))
+- **sendcloud:** stringify body ([3c520fc](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/3c520fc8a1c1c8f7cf632e2ca0833f3f4dd39878))
+- **sendcloud:** test fixes after using bodyparser rawbody ([3f6fe7d](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/3f6fe7d023f850b17797f56f80ad38ee2183410b))
+- **sendcloud:** trackingcodes ([3ace615](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/3ace615beaab6ad8479fd81dadd5b6933dd7603c))
+- **sendcloud:** unused addnavitem ([5fe759f](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/5fe759f1f4f77088292420e0e0803c9d812cc900))
+- wip ([e535566](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/e535566aba82da6ad8b36eef5f483678eb70943a))
+- wip sendcloud plugin ([ff4c0e5](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/ff4c0e5b04f123b63836fb89a24cf4de988c49c0))
+- wip sendcloud testing ([a997d80](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/a997d804d558adcaea3bd8292b47aa868b44da47))
+
+# [3.18.0](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v3.17.2...v3.18.0) (2022-09-26)
+
+### Features
+
+- **webhook:** pass injector to requestFn ([9c3a77e](https://github.com/martijnvdbrug/pinelab-vendure-plugins/commit/9c3a77ea6646157b1a1ab134bd79518bd6d4a9cb))
+
 ## [3.17.2](https://github.com/martijnvdbrug/pinelab-vendure-plugins/compare/v3.17.1...v3.17.2) (2022-09-19)
 
 ### Bug Fixes